+++ fourmi.h
@@ -30,10 +30,10 @@
 /** Données relative au parcourt d'une fourmi
  */
 typedef struct {
-    float L;            ///< Longueur d'un chemin, somme des longueurs de chaque arc constituant le chemin
+    double L;                    ///< Longueur d'un chemin, somme des longueurs de chaque arc constituant le chemin
     int nb_villes_deja_visite;
     bool parcourt_valide;
-    Ville *tabu[];      ///< Liste des villes déjà parcourues par la fourmi k. La plus plus récente est enregistré dans tabu[nb_villes_deja_visite - 1]; La taille de ce tableau est constante et est égale au nombre de villes dans la simulation;
+    Ville *tabu[];              ///< Liste des villes déjà parcourues par la fourmi k. La plus plus récente est enregistré dans tabu[nb_villes_deja_visite - 1]; La taille de ce tableau est constante et est égale au nombre de villes dans la simulation;
 } Fourmi;
 
 /** Initialise la structure fourmi
@@ -52,7 +52,7 @@ bool deja_visite(Ville *a_visiter, Ville *deja_visite[], int nb_villes_deja_visi
  *  \param proba_ville tableau aloué, non initialisé servant à calculer la la probabilité de chaque ville d'être tirée.
  *  \param deja_visite tableau contenant la liste des villes dejà visitées durant ce parcourt
  */
-void ville_suivante(Fourmi *f, int alpha, int beta, float proba_ville[], bool deja_visite[]);
+void ville_suivante(Fourmi *f, int alpha, int beta, double proba_ville[], bool deja_visite[]);
 
 /** Détermine le parcourt de la fourmie fourmie actuelle, valide ce parcourt puis met à jour le graph
  *  \param fourmi_actuelle Fourmi va pantounir le graph
@@ -66,7 +66,7 @@ void ville_suivante(Fourmi *f, int alpha, int beta, float proba_ville[], bool de
  *  \note Avant l'appel de cette fonction, le champ L (la distance du parcourt) de la meilleur fourmi doit etre initialisé
  *  \note Après l'appel de cette fonction, les données de fourmi_actuelle peuvent avoir changée et etre incohérente
  */
-void parcourt(Fourmi *fourmi_actuelle, Ville villes[], int nb_villes, bool ville_visitees[], int alpha, int beta, float proba_ville[]);
+void parcourt(Fourmi *fourmi_actuelle, Ville villes[], int nb_villes, bool ville_visitees[], int alpha, int beta, double proba_ville[]);
 
 /** Valide le parcourt d'une fourmi
  *  \param f Fourmi dont le parcourt est à tester
@@ -89,7 +89,7 @@ bool parcourt_valide(Fourmi *f, int nb_villes, bool ville_visitees[]);
  *  \note Avant l'appel de cette fonction, le champ L (la distance du parcourt) de la meilleur fourmi doit etre initialisé
  *  \note Après l'appel de cette fonction, les données de fourmi_actuelle peuvent avoir changée et etre incohérente
  */
-void parcourt_update(Fourmi **fourmi_actuelle, Fourmi **meilleure_fourmi, int nb_villes, bool ville_visitees[], float evaporation, float depot_pheromones);
+void parcourt_update(Fourmi **fourmi_actuelle, Fourmi **meilleure_fourmi, int nb_villes, bool ville_visitees[], double evaporation, double depot_pheromones);
 
 /** Affiche le parcourt d'une fourmie
  *  \note Avant de l'afficher, vérifie si le parcourt est valide
@@ -119,7 +119,7 @@ void affiche_parcourt(Fourmi *f, int nb_villes, bool ville_visitees[]);
 
  */
 void explore_graph(Ville villes[], Arc arcs[], Fourmi *(*fourmis[])
-    , Fourmi *meilleure_fourmi, bool ville_visitees[], float proba_ville[]
+    , Fourmi *meilleure_fourmi, bool ville_visitees[], double proba_ville[]
     , int nb_villes, int nb_fourmis
     , int max_cycle, double alpha, double beta, double evaporation, double depot_pheromones);
 #endif /* FOURMI_H */
